gradle plugin: Relocate build method to BundleTaskConvention
This avoids the need for a static method. Also updated logging to use
SLF4J parameters to avoid String conversion when log level not active.
Added check for errors after writing bundle in case any occurred. If
there are any Builder errors, the bundle file is deleted.
