+++ biz.aQute.bnd.gradle/src/aQute/bnd/gradle/BundleTaskConvention.groovy
@@ -19,17 +19,21 @@
 
 package aQute.bnd.gradle
 
+import aQute.bnd.osgi.Builder
+import aQute.bnd.osgi.Constants
+import aQute.bnd.osgi.Jar
+import aQute.bnd.version.Version
+import org.gradle.api.GradleException
 import org.gradle.api.Project
-import org.gradle.api.Task
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.tasks.SourceSet
 
 class BundleTaskConvention {
+  private final org.gradle.api.tasks.bundling.Jar task
   private File bndfile
   private Configuration configuration
   private SourceSet sourceSet
 
-  BundleTaskConvention(Task task) {
   /**
    * Create a BundleTaskConvention for the specified Jar task.
    *
@@ -37,6 +41,8 @@ class BundleTaskConvention {
    * This also sets the default values for the added properties
    * and adds the bnd file to the task inputs.
    */
+  BundleTaskConvention(org.gradle.api.tasks.bundling.Jar task) {
+    this.task = task
     def Project project = task.project
     // Set default property values
     setBndfile(project.file('bnd.bnd'))
@@ -48,42 +54,161 @@ class BundleTaskConvention {
     }
   }
 
-  File getBndfile() {
   /**
    * Get the bndfile property.
    */
+  public File getBndfile() {
     return bndfile
   }
-  void setBndfile(File bndfile) {
   /**
    * Set the bndfile property.
    */
+  public void setBndfile(File bndfile) {
     this.bndfile = bndfile
   }
 
-  Configuration getConfiguration() {
   /**
    * Get the configuration property.
    */
+  public Configuration getConfiguration() {
     return configuration
   }
-  void setConfiguration(Configuration configuration) {
   /**
    * Set the configuration property.
    */
+  public void setConfiguration(Configuration configuration) {
     this.configuration = configuration
   }
 
-  SourceSet getSourceSet() {
   /**
    * Get the sourceSet property.
    */
+  public SourceSet getSourceSet() {
     return sourceSet
   }
-  void setSourceSet(SourceSet sourceSet) {
   /**
    * Set the sourceSet property.
    */
+  public void setSourceSet(SourceSet sourceSet) {
     this.sourceSet = sourceSet
+  }
+
+  private void buildBundle() {
+    task.configure {
+      // create Builder and set trace level from gradle
+      def Builder builder = new Builder()
+      builder.setTrace(logger.isDebugEnabled())
+      try {
+        // load any task manifest entries into the builder
+        builder.addProperties(manifest.effectiveManifest.attributes)
+
+        // if the bnd file exists, set it as the builder properties
+        if (bndfile.isFile()) {
+          builder.setProperties(bndfile, project.projectDir)
+        } else {
+          builder.setBase(project.projectDir)
+        }
+
+        // If no bundle to be built, we have nothing to do
+        if (Builder.isTrue(builder.getProperty(Constants.NOBUNDLES))) {
+          return
+        }
+
+        // Reject sub-bundle projects
+        if ((builder.getSubBuilders().size() != 1) || ((builder.getPropertiesFile() != null) && !bndfile.equals(builder.getPropertiesFile()))) {
+          throw new GradleException('Sub-bundles are not supported by this task')
+        }
+
+        // copy Jar task generated jar to temporaryDir
+        project.copy {
+          from archivePath
+          into temporaryDir
+        }
+        def File jar = new File(temporaryDir, archiveName)
+
+        // set builder classpath
+        def Set<File> artifacts = new LinkedHashSet<File>(configuration.resolvedConfiguration.resolvedArtifacts*.file)
+        artifacts.add(jar)
+        builder.setClasspath(artifacts.toArray(new File[artifacts.size()]))
+        logger.debug 'builder classpath: {}', builder.getClasspath()*.getSource()
+
+        // set builder sourcepath
+        def Set<File> srcDirs = new LinkedHashSet<File>(sourceSet.java.srcDirs)
+        builder.setSourcepath(srcDirs.toArray(new File[srcDirs.size()]))
+        logger.debug 'builder sourcepath: {}', builder.getSourcePath()
+
+        // Include entire contents of Jar task generated jar except the manifest
+        def String includes = builder.getProperty(Constants.INCLUDERESOURCE)
+        if (isEmpty(includes)) {
+          includes = "@${jar}!/!META-INF/MANIFEST.MF"
+        } else {
+          includes = "@${jar}!/!META-INF/MANIFEST.MF,${includes}"
+        }
+        builder.setProperty(Constants.INCLUDERESOURCE, includes)
+
+        // set bundle symbolic name from tasks's baseName property if necessary
+        def String bundleSymbolicName = builder.getProperty(Constants.BUNDLE_SYMBOLICNAME)
+        if (isEmpty(bundleSymbolicName)) {
+          builder.setProperty(Constants.BUNDLE_SYMBOLICNAME, baseName)
+        }
+
+        // set bundle version from task's version if necessary
+        def String bundleVersion = builder.getProperty(Constants.BUNDLE_VERSION)
+        if (isEmpty(bundleVersion)) {
+          builder.setProperty(Constants.BUNDLE_VERSION, Version.parseVersion(version).toString())
+        }
+
+        logger.debug 'builder properties: {}', builder.getProperties()
+
+        // Build bundle
+        def Jar bundleJar = builder.build()
+        bundleJar.updateModified(jar.lastModified(), 'time of Jar task generated jar')
+        if (!builder.getErrors().isEmpty()) {
+          // if we already have an error; fail now
+          builder.getWarnings().each {
+            logger.warn 'Warning: {}', it
+          }
+          builder.getErrors().each {
+            logger.error 'Error  : {}', it
+          }
+          failBuild("Bundle ${archiveName} has errors")
+        }
+
+        // Write out the bundle
+        try {
+          bundleJar.write(archivePath)
+        } catch (Exception e) {
+          failBuild("Bundle ${archiveName} failed to build: ${e.getMessage()}", e)
+        } finally {
+          bundleJar.close()
+        }
+
+        builder.getWarnings().each {
+          logger.warn 'Warning: {}', it
+        }
+        builder.getErrors().each {
+          logger.error 'Error  : {}', it
+        }
+        if (!builder.getErrors().isEmpty()) {
+          failBuild("Bundle ${archiveName} has errors")
+        }
+      } finally {
+        builder.close()
+      }
+    }
+  }
+
+  private void failBuild(String msg) {
+    task.archivePath.delete()
+    throw new GradleException(msg)
+  }
+
+  private void failBuild(String msg, Exception e) {
+    task.archivePath.delete()
+    throw new GradleException(msg, e)
+  }
+
+  private boolean isEmpty(String header) {
+    return (header == null) || header.trim().isEmpty() || Constants.EMPTY_HEADER.equals(header)
   }
 }
